+++ model/service/dto/Client.go
@@ -4,5 +4,5 @@ import "time"
 
 type Client struct {
 	LastPayment time.Time
-	TotalAmount int64
+	TotalAmount float64
 }
